+++ src/services/pocket-relayer.ts
@@ -13,9 +13,10 @@ import { MetricsRecorder } from '../services/metrics-recorder'
 import { ConsensusFilterOptions, SyncChecker, SyncCheckOptions } from '../services/sync-checker'
 import { removeNodeFromSession } from '../utils/cache'
 import { MAX_RELAYS_ERROR } from '../utils/constants'
-import { checkSecretKey, SecretKeyDetails, updateConfiguration } from '../utils/pocket'
+import { checkEnforcementJSON, checkWhitelist, checkSecretKey, SecretKeyDetails } from '../utils/enforcements'
+import { parseMethod } from '../utils/parsing'
+import { updateConfiguration } from '../utils/pocket'
 import { filterCheckedNodes, isCheckPromiseResolved, loadBlockchain } from '../utils/relayer'
-import { checkEnforcementJSON, checkWhitelist, parseMethod } from '../utils/string'
 import { SendRelayOptions } from '../utils/types'
 import { enforceEVMLimits } from './limiter'
 const logger = require('../services/logger')
