+++ src/services/sync-checker.ts
@@ -5,7 +5,7 @@ import { MetricsRecorder } from '../services/metrics-recorder'
 import { blockHexToDecimal } from '../utils/block'
 import { removeNodeFromSession, getNodeNetworkData } from '../utils/cache'
 import { MAX_RELAYS_ERROR } from '../utils/constants'
-import { checkEnforcementJSON } from '../utils/string'
+import { checkEnforcementJSON } from '../utils/enforcements'
 
 const logger = require('../services/logger')
