+++ tests/unit/pocket-relayer.unit.ts
@@ -16,9 +16,10 @@ import { MetricsRecorder } from '../../src/services/metrics-recorder'
 import { PocketRelayer } from '../../src/services/pocket-relayer'
 import { ConsensusFilterOptions, SyncChecker, SyncCheckOptions } from '../../src/services/sync-checker'
 import { MAX_RELAYS_ERROR } from '../../src/utils/constants'
-import { checkSecretKey, updateConfiguration } from '../../src/utils/pocket'
+import { checkWhitelist, checkSecretKey } from '../../src/utils/enforcements'
+import { parseMethod } from '../../src/utils/parsing'
+import { updateConfiguration } from '../../src/utils/pocket'
 import { loadBlockchain } from '../../src/utils/relayer'
-import { checkWhitelist, parseMethod } from '../../src/utils/string'
 import { gatewayTestDB } from '../fixtures/test.datasource'
 import { metricsRecorderMock } from '../mocks/metricsRecorder'
 import { DEFAULT_NODES, PocketMock } from '../mocks/pocketjs'
